requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.6
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.1.0
Flask-Cors==4.0.0
frozenlist==1.4.0
functions-framework==3.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.4
packaging==23.1
pandas==2.2.2
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
requests==2.27.1
rsa==4.9
six==1.16.0
slack-bolt==1.17.0
slack-sdk==3.20.2
slackclient==2.9.4
tzdata==2024.1
urllib3==1.26.16
watchdog==3.0.0
Werkzeug==2.3.6
yarl==1.9.2
